Moved Nicety into its own file
